+++ CHANGELOG.md
@@ -9,6 +9,8 @@ methods) in the DB.
 
 ### Changed
 
+- concord-server: store `initiatorId` instead of `initiator` username
+in the process queue table;
 - project-model: fix handling of `null` values in `set` step.